+++ contracts/interfaces/IStorageDEX.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.21;
 interface IStorageDEX {
   function sessionManager() external view returns (address);
 
-  function withdrawMarginLevel() external view returns (int112);
+  function withdrawMarginLevel() external view returns (int256);
 }
